Remove unused stencilManager
The aim of this class was to save and restore the content of the stencil
buffer before and after using Goldfeather. The Windows implementation
using the ARB_buffer_region was working ok as far as I remember, however
for linux there was no such implementation. The extensionless implementation
never worked correctly. The implementation that was effectively always
used was the empty implementation which did not save and restore at all.

Removing the present code, in particular, removes a direct dependency on
WGL on Windows.

A good way to implement this with modern OpenGL could be glBlitFramebuffer.