Add ParityZPassOptimization switch
This is for the Goldfeather algorithm. The new default is that, to compute
the parity, GL_GREATER is used. This largely avoids the problems when
the camera is inside the CSG model and partly clipped by the view frustum,
which caused rendering errors then. The old setting GL_LEQUAL can be set
by setting ParityZPassOptimization on.

The old setting could be slightly faster, but my results were inconsistent.
This needs more investigation.