Fixed depth bounds test path when using the old z-less parity test
When discarding fragments, the depth bounds test must be not applied,
because we want to write depth values 1.0 for all discarded fragments.

Note that generally the depth bounds test seems to be questionable.
It is notoriously difficult to set the correct z values for the bounding
boxes of the provided primitives.