Renamed ParityZPassSetting to CameraOutsideOptimization
This prepares for the upcoming change in the SCS algorithm
which will allow for the camera to be inside of the CSG model.

Also changed the example application. In the Camera menu,
there are now 3 options

Camera outside (with performance optimizations)
Camera outside (without performance optimizations)
Camera inside

The Settings menu has been removed. It is no longer possible
to combine the "Camera inside" setting with the optimization
that the camera is outside, which would result in rendering errors.