Refined SCS change to allow camera inside the model
Again, for #13.

Can be disabled or enabled by CameraOutsideOptimizatio
Default of the optimization is current off, i.e., the slower,
but more compatible path is chosen.

Added documentation in opencsg.h

Refined the stencil operation for the z-less rendering of the
subtracted primitive.

Fixed the double occlusion query - I believe this is not defined.