Changed SCS wrt CameraOutsideOptimization default behavior
Only if this setting is turned off, the slower rendering path to
improve the behavior for the camera inside the model is chosen.
I decided for that because even this path, there are known, simple
rendering errors (the pipe in the example), so paying the slower
rendering performance is probably not worth.

Also enhance the example application accordingly. I don't like the new
menu, maybe the forth menu entry should be restored.